+++ stories/FButtonLink.stories.ts
@@ -1,7 +1,6 @@
-import { StoryFn } from '@storybook/vue3';
-
 import FButtonLink from '@/components/router/FButtonLink.vue';
 
+import type { StoryFn } from '@storybook/vue3';
 import type { FButtonLinkProps } from '@/components/router/FButtonLink.vue';
 
 export default {