+++ stories/FCreditCardInput.stories.ts
@@ -1,9 +1,9 @@
-import { StoryFn } from '@storybook/vue3';
 import { ref } from 'vue';
 
 import FCreditCardInput from '@/components/form/FCreditCardInput.vue';
 import FButton from '@/components/FButton.vue';
 
+import type { StoryFn } from '@storybook/vue3';
 import type { FCreditCardInputProps } from '@/components/form/FCreditCardInput.vue';
 
 export default {