+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.22.1 (November 27, 2023)
+
+### Fixes
+
+- [FPhoneInput] Properly handle form reset logic
+- [FPhoneInput] Bump `maska` to avoid sometimes buggy mask behavior
+- [Typing] Upgrade tsconfig and ensure components link
+- [FForm] Add story to document validation behavior
+- [Icons] Warn about missing icon only if name is not null
+- [FSelect] Remove sync watcher conflicting with form reset
+- [FLocaleSelect] Properly send name to FSelect to ensure from reset
+
 ## 0.22.0 (November 10, 2023)
 
 ### Features